hourly, daily, past bug fixes
A few bugs were fixed in this commit. Moment.js was not being called on
properly. I changed it to use the subtract method to calculate the date
range for the past week, month, and year. Previously it was using the
`days` method, which was incorrect.

Additionally, I had the threshold for getting hourly and daily sampling
too low. I thought the maximum sampling was set to record the temperature
every 90 seconds. It is now set to record the temperature every 180
seconds. getNewData.php was modified to get the first recording within
3 minutes after the hour to correct for this.